Fix the way INT is set within ToPA and the way the buffer size is computed

+++ IntelPTDriver/IntelPTDriver/Public.h
@@ -41,7 +41,7 @@ typedef struct _COMM_BUFFER_ADDRESS {
     
     PACKET_HEADER_INFORMATION Header;
 
-    unsigned BufferSize;
+    unsigned long BufferSize;
     union {
         PVOID BufferAddress;
         PVOID ImageBaseAddress;
